UCT/MM: Organize code in base/posix/sysv/xpmem subdirectories
